Fixed merge problems.
